Updates CodeQL workflow actions
Updates the CodeQL workflow to use the latest versions of the actions. This ensures that the workflow benefits from the latest features, bug fixes, and security updates provided by the action maintainers.
Specifically, updates the checkout, setup-node, init, autobuild, and analyze actions to their v4 or v3 counterparts, as appropriate, and node version to v22.